trying cmv

+++ Enzymes/cmv/dataset.py
@@ -126,7 +126,7 @@ def load(dataset):
             labels.append(np.array(list(nx.get_node_attributes(graph, 'label').values())))
             feat.append(np.array(list(nx.get_node_attributes(graph, 'feat').values())))
 
-        adj, diff, feat, labels = np.array(adj), np.array(diff), np.array(feat), np.array(labels)
+        adj, diff, feat, labels = np.array(adj), np.array(diff), np.array(feat), np.array(labels).astype(int)
 
         print('label_type ', print(labels.dtype))
